refactor: split sync event processor tests
