+++ test/features/sync/matrix/sync_event_processor_test.dart
@@ -2,26 +2,16 @@
 
 import 'dart:convert';
 import 'dart:io';
-import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/classes/entry_link.dart';
-import 'package:lotti/classes/entry_text.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/journal_update_result.dart';
 import 'package:lotti/database/logging_types.dart';
-import 'package:lotti/features/agents/model/agent_config.dart';
-import 'package:lotti/features/agents/model/agent_domain_entity.dart';
-import 'package:lotti/features/agents/model/agent_enums.dart';
-import 'package:lotti/features/agents/model/agent_link.dart';
-import 'package:lotti/features/agents/wake/wake_orchestrator.dart';
-import 'package:lotti/features/ai/model/ai_config.dart';
 import 'package:lotti/features/sync/matrix/pipeline/attachment_index.dart';
 import 'package:lotti/features/sync/matrix/sync_event_processor.dart';
 import 'package:lotti/features/sync/model/sync_message.dart';
 import 'package:lotti/features/sync/sequence/sync_sequence_log_service.dart';
-import 'package:lotti/features/sync/sequence/sync_sequence_payload_type.dart';
 import 'package:lotti/features/sync/vector_clock.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/db_notification.dart';
@@ -34,110 +24,11 @@ import 'package:path/path.dart' as path;
 import '../../../helpers/fallbacks.dart';
 import '../../../mocks/mocks.dart';
 import '../../../test_data/test_data.dart';
-
-class MockJournalEntityLoader extends Mock implements SyncJournalEntityLoader {}
+import 'sync_event_processor_test_helpers.dart';
 
 void main() {
-  setUpAll(() {
-    registerAllFallbackValues();
-    registerFallbackValue(
-      EntryLink.basic(
-        id: 'link-id',
-        fromId: 'from',
-        toId: 'to',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: null,
-      ),
-    );
-    registerFallbackValue(measurableWater);
-    registerFallbackValue(Uri.parse('mxc://placeholder'));
-    registerFallbackValue(Exception('test'));
-    registerFallbackValue(
-      const SyncBackfillRequest(entries: [], requesterId: ''),
-    );
-    registerFallbackValue(
-      const SyncBackfillResponse(hostId: '', counter: 0, deleted: false),
-    );
-    registerFallbackValue(const VectorClock({'fallback': 1}));
-    registerFallbackValue(SyncSequencePayloadType.journalEntity);
-  });
-  // Helper to normalize leading separators across platforms so that
-  // path.join(docDir, rel) never treats rel as absolute.
-  String stripLeadingSlashes(String s) =>
-      s.replaceFirst(RegExp(r'^[\\/]+'), '');
-
-  late MockEvent event;
-  late MockJournalDb journalDb;
-  late MockUpdateNotifications updateNotifications;
-  late MockLoggingService loggingService;
-  late MockAiConfigRepository aiConfigRepository;
-  late MockJournalEntityLoader journalEntityLoader;
-  late MockSettingsDb settingsDb;
-  late SyncEventProcessor processor;
-
-  setUp(() {
-    event = MockEvent();
-    journalDb = MockJournalDb();
-    updateNotifications = MockUpdateNotifications();
-    loggingService = MockLoggingService();
-    aiConfigRepository = MockAiConfigRepository();
-    journalEntityLoader = MockJournalEntityLoader();
-    settingsDb = MockSettingsDb();
-
-    when(
-      () => journalDb.updateJournalEntity(any<JournalEntity>()),
-    ).thenAnswer((_) async => JournalUpdateResult.applied());
-    when(
-      () => journalDb.entryLinkById(any<String>()),
-    ).thenAnswer((_) async => null);
-    when(
-      () => journalDb.upsertEntryLink(any<EntryLink>()),
-    ).thenAnswer((_) async => 1);
-    when(
-      () => journalDb.upsertEntityDefinition(any<EntityDefinition>()),
-    ).thenAnswer((_) async => 1);
-    when(
-      () => updateNotifications.notify(
-        any<Set<String>>(),
-        fromSync: any<bool>(named: 'fromSync'),
-      ),
-    ).thenAnswer((_) {});
-    stubLoggingService(loggingService);
-    when(
-      () => aiConfigRepository.saveConfig(
-        any<AiConfig>(),
-        fromSync: any<bool>(named: 'fromSync'),
-      ),
-    ).thenAnswer((_) async {});
-    when(
-      () => aiConfigRepository.deleteConfig(
-        any<String>(),
-        fromSync: any<bool>(named: 'fromSync'),
-      ),
-    ).thenAnswer((_) async {});
-
-    when(() => event.eventId).thenReturn('event-id');
-    when(() => event.originServerTs).thenReturn(DateTime(2024));
-
-    when(
-      () => settingsDb.itemByKey(any<String>()),
-    ).thenAnswer((_) async => null);
-    when(
-      () => settingsDb.saveSettingsItem(any<String>(), any<String>()),
-    ).thenAnswer((_) async => 1);
-
-    processor = SyncEventProcessor(
-      loggingService: loggingService,
-      updateNotifications: updateNotifications,
-      aiConfigRepository: aiConfigRepository,
-      settingsDb: settingsDb,
-      journalEntityLoader: journalEntityLoader,
-    );
-  });
-
-  String encodeMessage(SyncMessage message) =>
-      base64.encode(utf8.encode(json.encode(message.toJson())));
+  setUpAll(registerSyncProcessorFallbacks);
+  setUp(setUpProcessorMocks);
 
   test(
     'processes journal entities via loader and updates notifications',
@@ -553,1907 +444,6 @@ void main() {
 
     verify(() => aiConfigRepository.deleteConfig(id, fromSync: true)).called(1);
   });
-
-  group('SyncEventProcessor - Agent Entities and Links', () {
-    late MockAgentRepository mockAgentRepo;
-
-    setUp(() {
-      mockAgentRepo = MockAgentRepository();
-      when(() => mockAgentRepo.upsertEntity(any())).thenAnswer((_) async {});
-      when(() => mockAgentRepo.upsertLink(any())).thenAnswer((_) async {});
-      processor.agentRepository = mockAgentRepo;
-    });
-
-    test('processes agent identity entity', () async {
-      final entity = AgentDomainEntity.agent(
-        id: 'agent-1',
-        agentId: 'agent-1',
-        kind: 'task_agent',
-        displayName: 'Test Agent',
-        lifecycle: AgentLifecycle.active,
-        mode: AgentInteractionMode.autonomous,
-        allowedCategoryIds: const {'cat-1'},
-        currentStateId: 'state-1',
-        config: const AgentConfig(),
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-      verify(
-        () => loggingService.captureEvent(
-          any<Object>(that: contains('agentEntity')),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-      verify(
-        () => updateNotifications.notify(
-          {'agent-1', 'AGENT_CHANGED'},
-          fromSync: true,
-        ),
-      ).called(1);
-    });
-
-    test('processes agent state entity', () async {
-      final entity = AgentDomainEntity.agentState(
-        id: 'state-1',
-        agentId: 'agent-1',
-        revision: 5,
-        slots: const AgentSlots(),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-        wakeCounter: 42,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-    });
-
-    test('processes agent message entity', () async {
-      final entity = AgentDomainEntity.agentMessage(
-        id: 'msg-1',
-        agentId: 'agent-1',
-        threadId: 'thread-1',
-        kind: AgentMessageKind.thought,
-        createdAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-        metadata: const AgentMessageMetadata(),
-        tokensApprox: 100,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-    });
-
-    test('processes agent message payload entity', () async {
-      final entity = AgentDomainEntity.agentMessagePayload(
-        id: 'payload-1',
-        agentId: 'agent-1',
-        createdAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-        content: const {'text': 'hello world'},
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-    });
-
-    test('processes agent report entity', () async {
-      final entity = AgentDomainEntity.agentReport(
-        id: 'report-1',
-        agentId: 'agent-1',
-        scope: 'current',
-        createdAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-        content: 'Report content',
-        confidence: 0.9,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-    });
-
-    test('processes agent report head entity', () async {
-      final entity = AgentDomainEntity.agentReportHead(
-        id: 'head-1',
-        agentId: 'agent-1',
-        scope: 'current',
-        reportId: 'report-1',
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-    });
-
-    test(
-      'processes wakeTokenUsage entity with templateId in notification',
-      () async {
-        final entity = AgentDomainEntity.wakeTokenUsage(
-          id: 'usage-1',
-          agentId: 'agent-1',
-          runKey: 'run-1',
-          threadId: 'thread-1',
-          modelId: 'models/gemini-2.5-pro',
-          createdAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-          templateId: 'tpl-1',
-          inputTokens: 100,
-          outputTokens: 50,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verify(
-          () => updateNotifications.notify(
-            {'agent-1', 'tpl-1', 'AGENT_CHANGED'},
-            fromSync: true,
-          ),
-        ).called(1);
-      },
-    );
-
-    test('processes wakeTokenUsage entity without templateId', () async {
-      final entity = AgentDomainEntity.wakeTokenUsage(
-        id: 'usage-2',
-        agentId: 'agent-1',
-        runKey: 'run-2',
-        threadId: 'thread-2',
-        modelId: 'models/gemini-2.5-pro',
-        createdAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-        inputTokens: 200,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-      verify(
-        () => updateNotifications.notify(
-          {'agent-1', 'AGENT_CHANGED'},
-          fromSync: true,
-        ),
-      ).called(1);
-    });
-
-    test('processes agent link (basic)', () async {
-      final link = AgentLink.basic(
-        id: 'link-1',
-        fromId: 'agent-1',
-        toId: 'state-1',
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentLink(
-        agentLink: link,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verify(() => mockAgentRepo.upsertLink(link)).called(1);
-      verify(
-        () => loggingService.captureEvent(
-          any<Object>(that: contains('agentLink')),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-      verify(
-        () => updateNotifications.notify(
-          {'agent-1', 'state-1', 'AGENT_CHANGED'},
-          fromSync: true,
-        ),
-      ).called(1);
-    });
-
-    test(
-      'no-ops a legacy SyncAgentBundle envelope so the marker advances; '
-      'children recover via the per-entity / per-link backfill path',
-      () async {
-        final entity = AgentDomainEntity.agentState(
-          id: 'state-1',
-          agentId: 'agent-1',
-          revision: 5,
-          slots: const AgentSlots(),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: const VectorClock({'host-a': 1}),
-        );
-        final link = AgentLink.basic(
-          id: 'link-1',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: const VectorClock({'host-a': 2}),
-        );
-        final message = SyncMessage.agentBundle(
-          agentId: 'agent-1',
-          wakeRunKey: 'run-1',
-          originatingHostId: 'host-a',
-          entities: [
-            SyncMessage.agentEntity(
-                  agentEntity: entity,
-                  status: SyncEntryStatus.update,
-                )
-                as SyncAgentEntity,
-          ],
-          links: [
-            SyncMessage.agentLink(
-                  agentLink: link,
-                  status: SyncEntryStatus.update,
-                )
-                as SyncAgentLink,
-          ],
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        // Bundle is intentionally not applied — the agent repo is not
-        // touched, no AGENT_CHANGED notifications fire, and process()
-        // completes cleanly so the inbound queue marker advances. The
-        // sender already recorded each child under per-entity /
-        // per-link sequence-log entries, so backfill picks them up if
-        // they are missing locally.
-        verifyNever(() => mockAgentRepo.upsertEntity(any()));
-        verifyNever(() => mockAgentRepo.upsertLink(any()));
-        verifyNever(
-          () => updateNotifications.notify(
-            any<Set<String>>(),
-            fromSync: any<bool>(named: 'fromSync'),
-          ),
-        );
-      },
-    );
-
-    test(
-      'processes agent link variants (agentState, messagePrev, etc)',
-      () async {
-        final links = [
-          AgentLink.agentState(
-            id: 'link-2',
-            fromId: 'a',
-            toId: 'b',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          ),
-          AgentLink.messagePrev(
-            id: 'link-3',
-            fromId: 'a',
-            toId: 'b',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          ),
-          AgentLink.messagePayload(
-            id: 'link-4',
-            fromId: 'a',
-            toId: 'b',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          ),
-          AgentLink.toolEffect(
-            id: 'link-5',
-            fromId: 'a',
-            toId: 'b',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          ),
-          AgentLink.agentTask(
-            id: 'link-6',
-            fromId: 'a',
-            toId: 'b',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          ),
-        ];
-
-        for (final link in links) {
-          reset(mockAgentRepo);
-          when(() => mockAgentRepo.upsertLink(any())).thenAnswer((_) async {});
-
-          final message = SyncMessage.agentLink(
-            agentLink: link,
-            status: SyncEntryStatus.update,
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await processor.process(event: event, journalDb: journalDb);
-
-          verify(() => mockAgentRepo.upsertLink(link)).called(1);
-        }
-      },
-    );
-
-    test('skips agent entity when agentRepository is null', () async {
-      processor.agentRepository = null;
-
-      final entity = AgentDomainEntity.agent(
-        id: 'agent-1',
-        agentId: 'agent-1',
-        kind: 'task_agent',
-        displayName: 'Test Agent',
-        lifecycle: AgentLifecycle.active,
-        mode: AgentInteractionMode.autonomous,
-        allowedCategoryIds: const {},
-        currentStateId: 'state-1',
-        config: const AgentConfig(),
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verifyNever(() => mockAgentRepo.upsertEntity(any()));
-      verify(
-        () => loggingService.captureEvent(
-          any<Object>(that: contains('ignored')),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-    });
-
-    test('skips agent link when agentRepository is null', () async {
-      processor.agentRepository = null;
-
-      final link = AgentLink.basic(
-        id: 'link-1',
-        fromId: 'a',
-        toId: 'b',
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentLink(
-        agentLink: link,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      verifyNever(() => mockAgentRepo.upsertLink(any()));
-      verify(
-        () => loggingService.captureEvent(
-          any<Object>(that: contains('ignored')),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-    });
-
-    test('propagates repository error on agent entity upsert', () async {
-      when(
-        () => mockAgentRepo.upsertEntity(any()),
-      ).thenAnswer((_) async => throw Exception('db error'));
-
-      final entity = AgentDomainEntity.agent(
-        id: 'agent-1',
-        agentId: 'agent-1',
-        kind: 'task_agent',
-        displayName: 'Test Agent',
-        lifecycle: AgentLifecycle.active,
-        mode: AgentInteractionMode.autonomous,
-        allowedCategoryIds: const {},
-        currentStateId: 'state-1',
-        config: const AgentConfig(),
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentEntity(
-        agentEntity: entity,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      // The processor catches, logs, and rethrows
-      await expectLater(
-        () => processor.process(event: event, journalDb: journalDb),
-        throwsA(isA<Exception>()),
-      );
-
-      verify(
-        () => loggingService.captureException(
-          any<Object>(),
-          domain: any<String>(named: 'domain'),
-          subDomain: any<String>(named: 'subDomain'),
-          stackTrace: any<StackTrace>(named: 'stackTrace'),
-        ),
-      ).called(1);
-    });
-
-    test('propagates repository error on agent link upsert', () async {
-      when(
-        () => mockAgentRepo.upsertLink(any()),
-      ).thenAnswer((_) async => throw Exception('db error'));
-
-      final link = AgentLink.basic(
-        id: 'link-1',
-        fromId: 'a',
-        toId: 'b',
-        createdAt: DateTime(2024, 3, 15),
-        updatedAt: DateTime(2024, 3, 15),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.agentLink(
-        agentLink: link,
-        status: SyncEntryStatus.update,
-      );
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await expectLater(
-        () => processor.process(event: event, journalDb: journalDb),
-        throwsA(isA<Exception>()),
-      );
-
-      verify(
-        () => loggingService.captureException(
-          any<Object>(),
-          domain: any<String>(named: 'domain'),
-          subDomain: any<String>(named: 'subDomain'),
-          stackTrace: any<StackTrace>(named: 'stackTrace'),
-        ),
-      ).called(1);
-    });
-
-    group('agent entity sequence log recording', () {
-      late MockSyncSequenceLogService mockSeqService;
-      late MockAgentRepository mockAgentRepoSeq;
-
-      setUp(() {
-        mockSeqService = MockSyncSequenceLogService();
-        mockAgentRepoSeq = MockAgentRepository();
-        when(
-          () => mockAgentRepoSeq.upsertEntity(any()),
-        ).thenAnswer((_) async {});
-        when(() => mockAgentRepoSeq.upsertLink(any())).thenAnswer((_) async {});
-      });
-
-      test('records received agent entity in sequence log', () async {
-        const vc = VectorClock({'host-A': 10});
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-seq-1',
-          agentId: 'agent-seq-1',
-          kind: 'task_agent',
-          displayName: 'Seq Agent',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-A',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenAnswer((_) async => []);
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: 'agent-seq-1',
-            vectorClock: vc,
-            originatingHostId: 'host-A',
-            coveredVectorClocks: null,
-            payloadType: SyncSequencePayloadType.agentEntity,
-            jsonPath: any(named: 'jsonPath'),
-          ),
-        ).called(1);
-      });
-
-      test('logs gap detection for agent entity', () async {
-        const vc = VectorClock({'host-B': 20});
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-gap-1',
-          agentId: 'agent-gap-1',
-          kind: 'task_agent',
-          displayName: 'Gap Agent',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-B',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenAnswer(
-          (_) async => [(hostId: 'host-B', counter: 18)],
-        );
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => loggingService.captureEvent(
-            contains('apply.agentEntity.gapsDetected count=1'),
-            domain: LogDomains.sync,
-            subDomain: 'processor.gapDetection',
-          ),
-        ).called(1);
-      });
-
-      test('handles recordReceivedEntry exception for agent entity', () async {
-        const vc = VectorClock({'host-C': 5});
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-err-1',
-          agentId: 'agent-err-1',
-          kind: 'task_agent',
-          displayName: 'Err Agent',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-C',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenThrow(Exception('seq log error'));
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        // Entity should still be upserted despite seq log error
-        verify(() => mockAgentRepoSeq.upsertEntity(entity)).called(1);
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-
-      test('records received agent link in sequence log', () async {
-        const vc = VectorClock({'host-A': 15});
-        final link = AgentLink.basic(
-          id: 'link-seq-1',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentLink(
-          agentLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-A',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenAnswer((_) async => []);
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: 'link-seq-1',
-            vectorClock: vc,
-            originatingHostId: 'host-A',
-            coveredVectorClocks: null,
-            payloadType: SyncSequencePayloadType.agentLink,
-            jsonPath: any(named: 'jsonPath'),
-          ),
-        ).called(1);
-      });
-
-      test('logs gap detection for agent link', () async {
-        const vc = VectorClock({'host-D': 12});
-        final link = AgentLink.basic(
-          id: 'link-gap-1',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentLink(
-          agentLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-D',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenAnswer(
-          (_) async => [
-            (hostId: 'host-D', counter: 9),
-            (hostId: 'host-D', counter: 10),
-          ],
-        );
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => loggingService.captureEvent(
-            contains('apply.agentLink.gapsDetected count=2'),
-            domain: LogDomains.sync,
-            subDomain: 'processor.gapDetection',
-          ),
-        ).called(1);
-      });
-
-      test('handles recordReceivedEntry exception for agent link', () async {
-        const vc = VectorClock({'host-E': 7});
-        final link = AgentLink.basic(
-          id: 'link-err-1',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentLink(
-          agentLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-E',
-        );
-
-        when(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        ).thenThrow(Exception('seq log error link'));
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        // Link should still be upserted despite seq log error.
-        verify(() => mockAgentRepoSeq.upsertLink(link)).called(1);
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'SYNC_SEQUENCE',
-            subDomain: 'recordReceived',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-
-      test('skips sequence log when vectorClock is null', () async {
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-no-vc',
-          agentId: 'agent-no-vc',
-          kind: 'task_agent',
-          displayName: 'No VC',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-A',
-        );
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verifyNever(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        );
-      });
-
-      test('skips sequence log when originatingHostId is null', () async {
-        const vc = VectorClock({'host-A': 10});
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-no-host',
-          agentId: 'agent-no-host',
-          kind: 'task_agent',
-          displayName: 'No Host',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: vc,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-          // no originatingHostId
-        );
-
-        final proc = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSeqService,
-        )..agentRepository = mockAgentRepoSeq;
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        await proc.process(event: event, journalDb: journalDb);
-
-        verifyNever(
-          () => mockSeqService.recordReceivedEntry(
-            entryId: any(named: 'entryId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-            coveredVectorClocks: any(named: 'coveredVectorClocks'),
-            payloadType: any(named: 'payloadType'),
-          ),
-        );
-      });
-    });
-
-    group('lifecycle side-effects on incoming identity', () {
-      late MockWakeOrchestrator mockOrchestrator;
-
-      setUp(() {
-        mockOrchestrator = MockWakeOrchestrator();
-        processor.wakeOrchestrator = mockOrchestrator;
-        when(
-          () => mockOrchestrator.removeSubscriptions(any()),
-        ).thenReturn(null);
-      });
-
-      test('removes subscriptions when agent is dormant', () async {
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-dormant',
-          agentId: 'agent-dormant',
-          kind: 'task_agent',
-          displayName: 'Dormant Agent',
-          lifecycle: AgentLifecycle.dormant,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verify(
-          () => mockOrchestrator.removeSubscriptions('agent-dormant'),
-        ).called(1);
-      });
-
-      test('removes subscriptions when agent is destroyed', () async {
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-destroyed',
-          agentId: 'agent-destroyed',
-          kind: 'task_agent',
-          displayName: 'Destroyed Agent',
-          lifecycle: AgentLifecycle.destroyed,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verify(
-          () => mockOrchestrator.removeSubscriptions('agent-destroyed'),
-        ).called(1);
-      });
-
-      test('restores subscriptions for active task_agent', () async {
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-active',
-          agentId: 'agent-active',
-          kind: 'task_agent',
-          displayName: 'Active Agent',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final taskLink = AgentLink.basic(
-          id: 'link-1',
-          fromId: 'agent-active',
-          toId: 'task-42',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        when(
-          () => mockAgentRepo.getLinksFrom(
-            'agent-active',
-            type: 'agent_task',
-          ),
-        ).thenAnswer((_) async => [taskLink]);
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verifyNever(() => mockOrchestrator.removeSubscriptions(any()));
-        verify(
-          () => mockOrchestrator.addSubscription(
-            any(
-              that: isA<AgentSubscription>().having(
-                (s) => s.agentId,
-                'agentId',
-                'agent-active',
-              ),
-            ),
-          ),
-        ).called(1);
-      });
-
-      test('does NOT remove subscriptions for non-identity entities', () async {
-        final entity = AgentDomainEntity.agentState(
-          id: 'state-1',
-          agentId: 'agent-1',
-          revision: 5,
-          slots: const AgentSlots(),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verifyNever(() => mockOrchestrator.removeSubscriptions(any()));
-      });
-
-      test('safe when wakeOrchestrator is null', () async {
-        processor.wakeOrchestrator = null;
-
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-dormant',
-          agentId: 'agent-dormant',
-          kind: 'task_agent',
-          displayName: 'Dormant Agent',
-          lifecycle: AgentLifecycle.dormant,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentEntity(
-          agentEntity: entity,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        // Should not throw even though lifecycle is dormant.
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        verifyNever(() => mockOrchestrator.removeSubscriptions(any()));
-      });
-    });
-
-    group('subscription restoration on incoming agent link', () {
-      late MockWakeOrchestrator mockOrchestrator;
-
-      setUp(() {
-        mockOrchestrator = MockWakeOrchestrator();
-        processor.wakeOrchestrator = mockOrchestrator;
-      });
-
-      test(
-        'agent_task link for active task_agent restores subscription',
-        () async {
-          final activeAgent = AgentDomainEntity.agent(
-            id: 'agent-1',
-            agentId: 'agent-1',
-            kind: 'task_agent',
-            displayName: 'Active Agent',
-            lifecycle: AgentLifecycle.active,
-            mode: AgentInteractionMode.autonomous,
-            allowedCategoryIds: const {},
-            currentStateId: 'state-1',
-            config: const AgentConfig(),
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          );
-
-          when(
-            () => mockAgentRepo.getEntity('agent-1'),
-          ).thenAnswer((_) async => activeAgent);
-
-          final link = AgentLink.agentTask(
-            id: 'link-1',
-            fromId: 'agent-1',
-            toId: 'task-42',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          );
-
-          final message = SyncMessage.agentLink(
-            agentLink: link,
-            status: SyncEntryStatus.update,
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await processor.process(event: event, journalDb: journalDb);
-
-          verify(() => mockAgentRepo.upsertLink(link)).called(1);
-          verify(
-            () => mockOrchestrator.addSubscription(
-              any(
-                that: isA<AgentSubscription>()
-                    .having((s) => s.agentId, 'agentId', 'agent-1')
-                    .having(
-                      (s) => s.matchEntityIds,
-                      'matchEntityIds',
-                      {'task-42'},
-                    ),
-              ),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'agent_task link for dormant agent does NOT restore subscription',
-        () async {
-          final dormantAgent = AgentDomainEntity.agent(
-            id: 'agent-1',
-            agentId: 'agent-1',
-            kind: 'task_agent',
-            displayName: 'Dormant Agent',
-            lifecycle: AgentLifecycle.dormant,
-            mode: AgentInteractionMode.autonomous,
-            allowedCategoryIds: const {},
-            currentStateId: 'state-1',
-            config: const AgentConfig(),
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          );
-
-          when(
-            () => mockAgentRepo.getEntity('agent-1'),
-          ).thenAnswer((_) async => dormantAgent);
-
-          final link = AgentLink.agentTask(
-            id: 'link-1',
-            fromId: 'agent-1',
-            toId: 'task-42',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          );
-
-          final message = SyncMessage.agentLink(
-            agentLink: link,
-            status: SyncEntryStatus.update,
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await processor.process(event: event, journalDb: journalDb);
-
-          verify(() => mockAgentRepo.upsertLink(link)).called(1);
-          verifyNever(() => mockOrchestrator.addSubscription(any()));
-        },
-      );
-
-      test(
-        'soft-deleted agent_task link does NOT restore subscription',
-        () async {
-          final link = AgentLink.agentTask(
-            id: 'link-deleted',
-            fromId: 'agent-1',
-            toId: 'task-42',
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-            deletedAt: DateTime(2024, 3, 16),
-          );
-
-          final message = SyncMessage.agentLink(
-            agentLink: link,
-            status: SyncEntryStatus.update,
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await processor.process(event: event, journalDb: journalDb);
-
-          verify(() => mockAgentRepo.upsertLink(link)).called(1);
-          // Deleted link must not trigger agent lookup or subscription.
-          verifyNever(() => mockAgentRepo.getEntity(any()));
-          verifyNever(() => mockOrchestrator.addSubscription(any()));
-        },
-      );
-
-      test('non-agent_task link does NOT trigger subscription logic', () async {
-        final link = AgentLink.basic(
-          id: 'link-1',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.agentLink(
-          agentLink: link,
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertLink(link)).called(1);
-        // getEntity should not be called for non-agent_task links.
-        verifyNever(() => mockAgentRepo.getEntity(any()));
-        verifyNever(() => mockOrchestrator.addSubscription(any()));
-      });
-    });
-
-    group('descriptor-only resolution (jsonPath)', () {
-      late Directory tempDir;
-
-      setUp(() {
-        tempDir = Directory.systemTemp.createTempSync(
-          'agent_resolve_test',
-        );
-        if (getIt.isRegistered<Directory>()) {
-          getIt.unregister<Directory>();
-        }
-        getIt.registerSingleton<Directory>(tempDir);
-      });
-
-      tearDown(() async {
-        await getIt.reset();
-        if (tempDir.existsSync()) {
-          tempDir.deleteSync(recursive: true);
-        }
-      });
-
-      test('resolves agent entity from jsonPath on disk', () async {
-        final entity = AgentDomainEntity.agent(
-          id: 'agent-disk',
-          agentId: 'agent-disk',
-          kind: 'task_agent',
-          displayName: 'Disk Agent',
-          lifecycle: AgentLifecycle.active,
-          mode: AgentInteractionMode.autonomous,
-          allowedCategoryIds: const {},
-          currentStateId: 'state-1',
-          config: const AgentConfig(),
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        const relativePath = '/agent_entities/agent-disk.json';
-        final normalized = stripLeadingSlashes(relativePath);
-        final file = File(path.join(tempDir.path, normalized));
-        file.parent.createSync(recursive: true);
-        file.writeAsStringSync(jsonEncode(entity.toJson()));
-
-        const message = SyncMessage.agentEntity(
-          status: SyncEntryStatus.update,
-          jsonPath: relativePath,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-      });
-
-      test('resolves agent link from jsonPath on disk', () async {
-        final link = AgentLink.basic(
-          id: 'link-disk',
-          fromId: 'agent-1',
-          toId: 'state-1',
-          createdAt: DateTime(2024, 3, 15),
-          updatedAt: DateTime(2024, 3, 15),
-          vectorClock: null,
-        );
-
-        const relativePath = '/agent_links/link-disk.json';
-        final normalized = stripLeadingSlashes(relativePath);
-        final file = File(path.join(tempDir.path, normalized));
-        file.parent.createSync(recursive: true);
-        file.writeAsStringSync(jsonEncode(link.toJson()));
-
-        const message = SyncMessage.agentLink(
-          status: SyncEntryStatus.update,
-          jsonPath: relativePath,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verify(() => mockAgentRepo.upsertLink(link)).called(1);
-      });
-
-      test('skips agent entity with no entity and no jsonPath', () async {
-        const message = SyncMessage.agentEntity(
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertEntity(any()));
-        verify(
-          () => loggingService.captureEvent(
-            any<Object>(that: contains('no payload and no jsonPath')),
-            domain: LogDomains.sync,
-            subDomain: 'processor.resolve',
-          ),
-        ).called(1);
-      });
-
-      test('skips agent link with no link and no jsonPath', () async {
-        const message = SyncMessage.agentLink(
-          status: SyncEntryStatus.update,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertLink(any()));
-        verify(
-          () => loggingService.captureEvent(
-            any<Object>(that: contains('no payload and no jsonPath')),
-            domain: LogDomains.sync,
-            subDomain: 'processor.resolve',
-          ),
-        ).called(1);
-      });
-
-      test('skips agent entity with path-traversal jsonPath', () async {
-        const message = SyncMessage.agentEntity(
-          status: SyncEntryStatus.update,
-          jsonPath: '../../../etc/passwd',
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertEntity(any()));
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'AGENT_SYNC',
-            subDomain: 'resolve.agentEntity.invalidPath',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-
-      test('skips agent link with path-traversal jsonPath', () async {
-        const message = SyncMessage.agentLink(
-          status: SyncEntryStatus.update,
-          jsonPath: '../../../etc/passwd',
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertLink(any()));
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'AGENT_SYNC',
-            subDomain: 'resolve.agentLink.invalidPath',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-
-      test(
-        'rethrows FileSystemException for missing agent entity file',
-        () async {
-          const message = SyncMessage.agentEntity(
-            status: SyncEntryStatus.update,
-            jsonPath: '/agent_entities/missing.json',
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await expectLater(
-            () => processor.process(event: event, journalDb: journalDb),
-            throwsA(isA<FileSystemException>()),
-          );
-        },
-      );
-
-      test(
-        'rethrows FileSystemException for missing agent link file',
-        () async {
-          const message = SyncMessage.agentLink(
-            status: SyncEntryStatus.update,
-            jsonPath: '/agent_links/missing.json',
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await expectLater(
-            () => processor.process(event: event, journalDb: journalDb),
-            throwsA(isA<FileSystemException>()),
-          );
-        },
-      );
-
-      test('skips agent entity with corrupt JSON file', () async {
-        const relativePath = '/agent_entities/corrupt.json';
-        final normalized = stripLeadingSlashes(relativePath);
-        final file = File(path.join(tempDir.path, normalized));
-        file.parent.createSync(recursive: true);
-        file.writeAsStringSync('not valid json {{{');
-
-        const message = SyncMessage.agentEntity(
-          status: SyncEntryStatus.update,
-          jsonPath: relativePath,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertEntity(any()));
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'AGENT_SYNC',
-            subDomain: 'resolve.agentEntity',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-
-      group('descriptor-based resolution (AttachmentIndex)', () {
-        late AttachmentIndex attachmentIndex;
-        late SyncEventProcessor processorWithIndex;
-        late MockEvent descriptorEvent;
-
-        setUp(() {
-          attachmentIndex = AttachmentIndex(logging: loggingService);
-          processorWithIndex = SyncEventProcessor(
-            loggingService: loggingService,
-            updateNotifications: updateNotifications,
-            aiConfigRepository: aiConfigRepository,
-            settingsDb: settingsDb,
-            journalEntityLoader: journalEntityLoader,
-            attachmentIndex: attachmentIndex,
-          );
-          processorWithIndex.agentRepository = mockAgentRepo;
-
-          descriptorEvent = MockEvent();
-          when(() => descriptorEvent.eventId).thenReturn('desc-event-id');
-          when(
-            () => descriptorEvent.attachmentMimetype,
-          ).thenReturn('application/json');
-          when(() => descriptorEvent.content).thenReturn({
-            'relativePath': '/agent_entities/agent-desc.json',
-          });
-        });
-
-        test(
-          'fetches agent entity from descriptor when file is missing',
-          () async {
-            final entity = AgentDomainEntity.agent(
-              id: 'agent-desc',
-              agentId: 'agent-desc',
-              kind: 'task_agent',
-              displayName: 'Descriptor Agent',
-              lifecycle: AgentLifecycle.active,
-              mode: AgentInteractionMode.autonomous,
-              allowedCategoryIds: const {},
-              currentStateId: 'state-1',
-              config: const AgentConfig(),
-              createdAt: DateTime(2024, 3, 15),
-              updatedAt: DateTime(2024, 3, 15),
-              vectorClock: null,
-            );
-
-            final bytes = Uint8List.fromList(
-              utf8.encode(jsonEncode(entity.toJson())),
-            );
-            when(
-              descriptorEvent.downloadAndDecryptAttachment,
-            ).thenAnswer((_) async => MatrixFile(bytes: bytes, name: 'e.json'));
-
-            attachmentIndex.record(descriptorEvent);
-
-            const message = SyncMessage.agentEntity(
-              status: SyncEntryStatus.update,
-              jsonPath: '/agent_entities/agent-desc.json',
-            );
-            when(() => event.text).thenReturn(encodeMessage(message));
-
-            await processorWithIndex.process(
-              event: event,
-              journalDb: journalDb,
-            );
-
-            verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-            verify(descriptorEvent.downloadAndDecryptAttachment).called(1);
-          },
-        );
-
-        test(
-          'fetches agent link from descriptor when file is missing',
-          () async {
-            final link = AgentLink.basic(
-              id: 'link-desc',
-              fromId: 'agent-1',
-              toId: 'state-1',
-              createdAt: DateTime(2024, 3, 15),
-              updatedAt: DateTime(2024, 3, 15),
-              vectorClock: null,
-            );
-
-            final bytes = Uint8List.fromList(
-              utf8.encode(jsonEncode(link.toJson())),
-            );
-
-            final linkDescriptorEvent = MockEvent();
-            when(
-              () => linkDescriptorEvent.eventId,
-            ).thenReturn('link-desc-event-id');
-            when(
-              () => linkDescriptorEvent.attachmentMimetype,
-            ).thenReturn('application/json');
-            when(() => linkDescriptorEvent.content).thenReturn({
-              'relativePath': '/agent_links/link-desc.json',
-            });
-            when(
-              linkDescriptorEvent.downloadAndDecryptAttachment,
-            ).thenAnswer((_) async => MatrixFile(bytes: bytes, name: 'l.json'));
-
-            attachmentIndex.record(linkDescriptorEvent);
-
-            const message = SyncMessage.agentLink(
-              status: SyncEntryStatus.update,
-              jsonPath: '/agent_links/link-desc.json',
-            );
-            when(() => event.text).thenReturn(encodeMessage(message));
-
-            await processorWithIndex.process(
-              event: event,
-              journalDb: journalDb,
-            );
-
-            verify(() => mockAgentRepo.upsertLink(link)).called(1);
-            verify(linkDescriptorEvent.downloadAndDecryptAttachment).called(1);
-          },
-        );
-
-        test(
-          'throws when descriptor download fails (no stale fallback)',
-          () async {
-            when(
-              descriptorEvent.downloadAndDecryptAttachment,
-            ).thenThrow(Exception('download failed'));
-
-            attachmentIndex.record(descriptorEvent);
-
-            const message = SyncMessage.agentEntity(
-              status: SyncEntryStatus.update,
-              jsonPath: '/agent_entities/agent-desc.json',
-            );
-            when(() => event.text).thenReturn(encodeMessage(message));
-
-            await expectLater(
-              () => processorWithIndex.process(
-                event: event,
-                journalDb: journalDb,
-              ),
-              throwsA(isA<FileSystemException>()),
-            );
-
-            verifyNever(() => mockAgentRepo.upsertEntity(any()));
-          },
-        );
-
-        test('throws when descriptor returns empty bytes', () async {
-          when(descriptorEvent.downloadAndDecryptAttachment).thenAnswer(
-            (_) async => MatrixFile(bytes: Uint8List(0), name: 'e.json'),
-          );
-
-          attachmentIndex.record(descriptorEvent);
-
-          const message = SyncMessage.agentEntity(
-            status: SyncEntryStatus.update,
-            jsonPath: '/agent_entities/agent-desc.json',
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await expectLater(
-            () => processorWithIndex.process(
-              event: event,
-              journalDb: journalDb,
-            ),
-            throwsA(isA<FileSystemException>()),
-          );
-
-          verifyNever(() => mockAgentRepo.upsertEntity(any()));
-        });
-
-        test('falls back to disk when no descriptor in index', () async {
-          final entity = AgentDomainEntity.agent(
-            id: 'agent-disk-fb',
-            agentId: 'agent-disk-fb',
-            kind: 'task_agent',
-            displayName: 'Disk Fallback',
-            lifecycle: AgentLifecycle.active,
-            mode: AgentInteractionMode.autonomous,
-            allowedCategoryIds: const {},
-            currentStateId: 'state-1',
-            config: const AgentConfig(),
-            createdAt: DateTime(2024, 3, 15),
-            updatedAt: DateTime(2024, 3, 15),
-            vectorClock: null,
-          );
-
-          const relativePath = '/agent_entities/agent-disk-fb.json';
-          final normalized = stripLeadingSlashes(relativePath);
-          final file = File(path.join(tempDir.path, normalized));
-          file.parent.createSync(recursive: true);
-          file.writeAsStringSync(jsonEncode(entity.toJson()));
-
-          const message = SyncMessage.agentEntity(
-            status: SyncEntryStatus.update,
-            jsonPath: relativePath,
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          await processorWithIndex.process(
-            event: event,
-            journalDb: journalDb,
-          );
-
-          verify(() => mockAgentRepo.upsertEntity(entity)).called(1);
-        });
-
-        test('skips agent entity with corrupt descriptor JSON', () async {
-          final bytes = Uint8List.fromList(utf8.encode('not valid json'));
-          when(
-            descriptorEvent.downloadAndDecryptAttachment,
-          ).thenAnswer((_) async => MatrixFile(bytes: bytes, name: 'e.json'));
-
-          attachmentIndex.record(descriptorEvent);
-
-          const message = SyncMessage.agentEntity(
-            status: SyncEntryStatus.update,
-            jsonPath: '/agent_entities/agent-desc.json',
-          );
-          when(() => event.text).thenReturn(encodeMessage(message));
-
-          // Descriptor fetched successfully but JSON parse fails →
-          // permanent skip (returns null), not a retry.
-          await processorWithIndex.process(
-            event: event,
-            journalDb: journalDb,
-          );
-
-          verifyNever(() => mockAgentRepo.upsertEntity(any()));
-          verify(
-            () => loggingService.captureException(
-              any<Object>(),
-              domain: 'AGENT_SYNC',
-              subDomain: 'resolve.agentEntity.parseFetched',
-              stackTrace: any<StackTrace>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        });
-      });
-
-      test('skips agent link with corrupt JSON file', () async {
-        const relativePath = '/agent_links/corrupt.json';
-        final normalized = stripLeadingSlashes(relativePath);
-        final file = File(path.join(tempDir.path, normalized));
-        file.parent.createSync(recursive: true);
-        file.writeAsStringSync('not valid json {{{');
-
-        const message = SyncMessage.agentLink(
-          status: SyncEntryStatus.update,
-          jsonPath: relativePath,
-        );
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        verifyNever(() => mockAgentRepo.upsertLink(any()));
-        verify(
-          () => loggingService.captureException(
-            any<Object>(),
-            domain: 'AGENT_SYNC',
-            subDomain: 'resolve.agentLink',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      });
-    });
-  });
-
-  test('logs exceptions for invalid base64 payloads', () async {
-    when(() => event.text).thenReturn('not-base64');
-
-    await expectLater(
-      processor.process(event: event, journalDb: journalDb),
-      throwsA(isA<FormatException>()),
-    );
-
-    verify(
-      () => loggingService.captureException(
-        any<Object>(),
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SyncEventProcessor',
-        stackTrace: any<StackTrace>(named: 'stackTrace'),
-      ),
-    ).called(1);
-  });
-
-  test('logs exceptions thrown by handlers', () async {
-    const message = SyncMessage.journalEntity(
-      id: 'entity-id',
-      jsonPath: '/entity.json',
-      vectorClock: null,
-      status: SyncEntryStatus.initial,
-    );
-    when(() => event.text).thenReturn(encodeMessage(message));
-    when(
-      () => journalEntityLoader.load(
-        jsonPath: '/entity.json',
-      ),
-    ).thenThrow(Exception('load failed'));
-
-    await expectLater(
-      processor.process(event: event, journalDb: journalDb),
-      throwsA(isA<Exception>()),
-    );
-
-    verify(
-      () => loggingService.captureException(
-        any<Object>(),
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SyncEventProcessor',
-        stackTrace: any<StackTrace>(named: 'stackTrace'),
-      ),
-    ).called(1);
-  });
-
-  test('skips message with unknown enum value (ArgumentError)', () async {
-    // Simulate a SyncMessage JSON with an unknown enum value that would
-    // cause $enumDecode to throw ArgumentError.
-    final badJson = {
-      'runtimeType': 'journalEntity',
-      'id': 'entity-id',
-      'jsonPath': '/entity.json',
-      'vectorClock': null,
-      'status': 'unknownEnumValue',
-    };
-    final encoded = base64.encode(utf8.encode(json.encode(badJson)));
-    when(() => event.text).thenReturn(encoded);
-
-    // Should NOT throw — the error is caught and logged.
-    await processor.process(event: event, journalDb: journalDb);
-
-    verify(
-      () => loggingService.captureEvent(
-        any<Object>(
-          that: contains('skipping undeserializable sync message'),
-        ),
-        domain: LogDomains.sync,
-        subDomain: 'processor.skipUnrecoverable',
-      ),
-    ).called(1);
-  });
-
-  test('skips message with FormatException from malformed JSON', () async {
-    // Valid base64/JSON but with a structure that causes FormatException
-    // when SyncMessage.fromJson tries to parse sub-fields.
-    final badJson = {
-      'runtimeType': 'journalEntity',
-      'id': 123, // wrong type — id should be String
-      'jsonPath': '/entity.json',
-      'vectorClock': null,
-      'status': 'initial',
-    };
-    final encoded = base64.encode(utf8.encode(json.encode(badJson)));
-    when(() => event.text).thenReturn(encoded);
-
-    // The error might be TypeError or similar — either way, if it's not
-    // ArgumentError or FormatException, it will rethrow through the outer
-    // catch. Let's verify it doesn't crash with an unrecoverable retry.
-    try {
-      await processor.process(event: event, journalDb: journalDb);
-      // If it didn't throw, it was caught as deserialization error → good.
-    } on Object {
-      // If it threw, it's a non-deserialization error that rethrows → also ok,
-      // but verify the outer catch logged it.
-      verify(
-        () => loggingService.captureException(
-          any<Object>(),
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SyncEventProcessor',
-          stackTrace: any<StackTrace>(named: 'stackTrace'),
-        ),
-      ).called(1);
-    }
-  });
-
-  test('skips message with CheckedFromJsonException from empty JSON', () async {
-    // An empty JSON object hits the default case in the generated
-    // _$SyncMessageFromJson switch, which throws CheckedFromJsonException.
-    // The processor catches this and logs a skip.
-    final badJson = <String, dynamic>{};
-    final encoded = base64.encode(utf8.encode(json.encode(badJson)));
-    when(() => event.text).thenReturn(encoded);
-
-    await processor.process(event: event, journalDb: journalDb);
-
-    verify(
-      () => loggingService.captureEvent(
-        any<String>(
-          that: contains('skipping undeserializable sync message'),
-        ),
-        domain: LogDomains.sync,
-        subDomain: 'processor.skipUnrecoverable',
-      ),
-    ).called(1);
-  });
-
   group('SyncEventProcessor listener -', () {
     test('cachePurgeListener with non-smart loader does not crash', () {
       final processorWithFileLoader = SyncEventProcessor(
@@ -3058,368 +1048,6 @@ void main() {
     });
   });
 
-  group('SyncEventProcessor - Embedded Entry Links', () {
-    test(
-      'processes embedded links after successful journal entity update',
-      () async {
-        final link1 = EntryLink.basic(
-          id: 'link-1',
-          fromId: 'entry-id',
-          toId: 'category-1',
-          createdAt: DateTime(2025, 1, 1),
-          updatedAt: DateTime(2025, 1, 1),
-          vectorClock: null,
-        );
-        final link2 = EntryLink.basic(
-          id: 'link-2',
-          fromId: 'entry-id',
-          toId: 'category-2',
-          createdAt: DateTime(2025, 1, 1),
-          updatedAt: DateTime(2025, 1, 1),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.journalEntity(
-          id: 'entry-id',
-          jsonPath: '/entry.json',
-          vectorClock: null,
-          status: SyncEntryStatus.initial,
-          entryLinks: [link1, link2],
-        );
-
-        when(
-          () => journalEntityLoader.load(jsonPath: '/entry.json'),
-        ).thenAnswer((_) async => fallbackJournalEntity);
-        when(() => event.text).thenReturn(encodeMessage(message));
-        when(() => journalDb.upsertEntryLink(link1)).thenAnswer((_) async => 1);
-        when(() => journalDb.upsertEntryLink(link2)).thenAnswer((_) async => 1);
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        // Verify both links were upserted
-        verify(() => journalDb.upsertEntryLink(link1)).called(1);
-        verify(() => journalDb.upsertEntryLink(link2)).called(1);
-
-        // Verify logging for each embedded link
-        verify(
-          () => loggingService.captureEvent(
-            contains(
-              'apply entryLink.embedded from=${link1.fromId} to=${link1.toId}',
-            ),
-            domain: LogDomains.sync,
-            subDomain: 'processor.apply.entryLink.embedded',
-          ),
-        ).called(1);
-
-        verify(
-          () => loggingService.captureEvent(
-            contains(
-              'apply entryLink.embedded from=${link2.fromId} to=${link2.toId}',
-            ),
-            domain: LogDomains.sync,
-            subDomain: 'processor.apply.entryLink.embedded',
-          ),
-        ).called(1);
-
-        // Verify summary log includes embedded links count
-        verify(
-          () => loggingService.captureEvent(
-            contains('embeddedLinks=2/2'),
-            domain: LogDomains.sync,
-            subDomain: 'processor.apply',
-          ),
-        ).called(1);
-
-        // Verify notifications sent for all affected IDs from both links
-        verify(
-          () => updateNotifications.notify(
-            {link1.fromId, link1.toId, link2.toId},
-            fromSync: true,
-          ),
-        ).called(1);
-      },
-    );
-
-    test('processes embedded links even when entity update is skipped', () async {
-      final link = EntryLink.basic(
-        id: 'link-1',
-        fromId: 'entry-id',
-        toId: 'category-1',
-        createdAt: DateTime(2025, 1, 1),
-        updatedAt: DateTime(2025, 1, 1),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.journalEntity(
-        id: 'entry-id',
-        jsonPath: '/entry.json',
-        vectorClock: const VectorClock({'old': 1}),
-        status: SyncEntryStatus.initial,
-        entryLinks: [link],
-      );
-
-      // Create an entry with newer vector clock so update is skipped
-      final newerEntry = JournalEntry(
-        meta: Metadata(
-          id: 'entry-id',
-          createdAt: DateTime(2025, 1, 1),
-          updatedAt: DateTime(2025, 1, 1),
-          dateFrom: DateTime(2025, 1, 1),
-          dateTo: DateTime(2025, 1, 1),
-          vectorClock: const VectorClock({'new': 2}),
-        ),
-        entryText: const EntryText(plainText: 'newer'),
-      );
-
-      when(
-        () => journalEntityLoader.load(
-          jsonPath: '/entry.json',
-          incomingVectorClock: const VectorClock({'old': 1}),
-        ),
-      ).thenAnswer((_) async => fallbackJournalEntity);
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(
-        () => journalDb.journalEntityById('entry-id'),
-      ).thenAnswer((_) async => newerEntry);
-      when(
-        () => journalDb.updateJournalEntity(any<JournalEntity>()),
-      ).thenAnswer(
-        (_) async => JournalUpdateResult.skipped(
-          reason: JournalUpdateSkipReason.olderOrEqual,
-        ),
-      );
-      when(() => journalDb.upsertEntryLink(link)).thenAnswer((_) async => 1);
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      // Verify link WAS upserted even though entity update was skipped.
-      // EntryLinks have their own vector clock for conflict resolution,
-      // so they should be processed regardless of journal entity status.
-      // This prevents gray calendar entries that rely on links for color lookup.
-      verify(() => journalDb.upsertEntryLink(link)).called(1);
-
-      // Verify logging for embedded link processing
-      verify(
-        () => loggingService.captureEvent(
-          contains(
-            'apply entryLink.embedded from=${link.fromId} to=${link.toId}',
-          ),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply.entryLink.embedded',
-        ),
-      ).called(1);
-
-      // Verify summary shows 1 embedded link processed
-      verify(
-        () => loggingService.captureEvent(
-          contains('embeddedLinks=1/1'),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-
-      // Verify notification sent for affected IDs from link
-      verify(
-        () => updateNotifications.notify(
-          {link.fromId, link.toId},
-          fromSync: true,
-        ),
-      ).called(1);
-    });
-
-    test(
-      'embedded link failure inside the journal-entity apply path aborts '
-      'the transaction so the entity row never lands without its links — '
-      'redelivery routes the same event to the duplicate path which does '
-      'NOT retry link upserts, so a swallowed failure here would leave '
-      'permanently missing edges',
-      () async {
-        final link1 = EntryLink.basic(
-          id: 'link-1',
-          fromId: 'entry-id',
-          toId: 'category-1',
-          createdAt: DateTime(2025, 1, 1),
-          updatedAt: DateTime(2025, 1, 1),
-          vectorClock: null,
-        );
-        final link2 = EntryLink.basic(
-          id: 'link-2',
-          fromId: 'entry-id',
-          toId: 'category-2',
-          createdAt: DateTime(2025, 1, 1),
-          updatedAt: DateTime(2025, 1, 1),
-          vectorClock: null,
-        );
-
-        final message = SyncMessage.journalEntity(
-          id: 'entry-id',
-          jsonPath: '/entry.json',
-          vectorClock: null,
-          status: SyncEntryStatus.initial,
-          entryLinks: [link1, link2],
-        );
-
-        when(
-          () => journalEntityLoader.load(jsonPath: '/entry.json'),
-        ).thenAnswer((_) async => fallbackJournalEntity);
-        when(() => event.text).thenReturn(encodeMessage(message));
-
-        // First link fails, second would succeed — but the loop must
-        // bail on the first failure and let the transaction roll back.
-        when(
-          () => journalDb.upsertEntryLink(link1),
-        ).thenThrow(Exception('Database error'));
-        when(() => journalDb.upsertEntryLink(link2)).thenAnswer((_) async => 1);
-
-        await expectLater(
-          processor.process(event: event, journalDb: journalDb),
-          throwsA(isA<Exception>()),
-        );
-
-        // The first link was attempted (and threw); the second was
-        // NOT — the loop exits on rethrow.
-        verify(() => journalDb.upsertEntryLink(link1)).called(1);
-        verifyNever(() => journalDb.upsertEntryLink(link2));
-
-        // The link-level capture still ran before the rethrow so the
-        // failure shows up in domain logs.
-        verify(
-          () => loggingService.captureException(
-            any<Exception>(),
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'apply.entryLink.embedded',
-            stackTrace: any<StackTrace>(named: 'stackTrace'),
-          ),
-        ).called(1);
-      },
-    );
-
-    test('processes empty embedded links list', () async {
-      const message = SyncMessage.journalEntity(
-        id: 'entry-id',
-        jsonPath: '/entry.json',
-        vectorClock: null,
-        status: SyncEntryStatus.initial,
-        entryLinks: [],
-      );
-
-      when(
-        () => journalEntityLoader.load(jsonPath: '/entry.json'),
-      ).thenAnswer((_) async => fallbackJournalEntity);
-      when(() => event.text).thenReturn(encodeMessage(message));
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      // Verify no links were upserted
-      verifyNever(() => journalDb.upsertEntryLink(any()));
-
-      // Verify summary shows 0/0 embedded links
-      verify(
-        () => loggingService.captureEvent(
-          contains('embeddedLinks=0/0'),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-    });
-
-    test('skips link processing when linkRows is 0 (no-op upsert)', () async {
-      final link = EntryLink.basic(
-        id: 'link-1',
-        fromId: 'entry-id',
-        toId: 'category-1',
-        createdAt: DateTime(2025, 1, 1),
-        updatedAt: DateTime(2025, 1, 1),
-        vectorClock: null,
-      );
-
-      final message = SyncMessage.journalEntity(
-        id: 'entry-id',
-        jsonPath: '/entry.json',
-        vectorClock: null,
-        status: SyncEntryStatus.initial,
-        entryLinks: [link],
-      );
-
-      when(
-        () => journalEntityLoader.load(jsonPath: '/entry.json'),
-      ).thenAnswer((_) async => fallbackJournalEntity);
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(() => journalDb.upsertEntryLink(link)).thenAnswer((_) async => 0);
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      // Verify link upsert was attempted
-      verify(() => journalDb.upsertEntryLink(link)).called(1);
-
-      // Verify no log for link application (rows was 0)
-      verifyNever(
-        () => loggingService.captureEvent(
-          contains('apply entryLink.embedded'),
-          domain: any(named: 'domain'),
-          subDomain: 'processor.apply.entryLink.embedded',
-        ),
-      );
-
-      // Verify summary shows 0 processed (since linkRows was 0)
-      verify(
-        () => loggingService.captureEvent(
-          contains('embeddedLinks=0/1'),
-          domain: LogDomains.sync,
-          subDomain: 'processor.apply',
-        ),
-      ).called(1);
-
-      // Verify notifications were still sent for affected IDs
-      verify(
-        () => updateNotifications.notify(
-          {link.fromId, link.toId},
-          fromSync: true,
-        ),
-      ).called(1);
-    });
-
-    test('syncs collapsed state from remote embedded link', () async {
-      final incomingLink = EntryLink.basic(
-        id: 'link-collapsed',
-        fromId: 'entry-id',
-        toId: 'category-1',
-        createdAt: DateTime(2025, 1, 1),
-        updatedAt: DateTime(2025, 1, 2),
-        vectorClock: null,
-        collapsed: true,
-      );
-
-      final message = SyncMessage.journalEntity(
-        id: 'entry-id',
-        jsonPath: '/entry.json',
-        vectorClock: null,
-        status: SyncEntryStatus.initial,
-        entryLinks: [incomingLink],
-      );
-
-      when(
-        () => journalEntityLoader.load(jsonPath: '/entry.json'),
-      ).thenAnswer((_) async => fallbackJournalEntity);
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(
-        () => journalDb.upsertEntryLink(any<EntryLink>()),
-      ).thenAnswer((_) async => 1);
-
-      await processor.process(event: event, journalDb: journalDb);
-
-      // Verify the upserted link has collapsed=true from the incoming link
-      final capturedLink =
-          verify(
-                () => journalDb.upsertEntryLink(captureAny<EntryLink>()),
-              ).captured.single
-              as EntryLink;
-      expect(capturedLink.id, 'link-collapsed');
-      expect(capturedLink.collapsed, isTrue);
-      expect(capturedLink.updatedAt, DateTime(2025, 1, 2));
-    });
-  });
-
   group('SyncEventProcessor - Backfill Messages', () {
     test('SyncBackfillRequest throws when no handler configured', () async {
       const message = SyncBackfillRequest(
@@ -3440,7 +1068,9 @@ void main() {
           isA<Error>().having(
             (e) => e.toString(),
             'toString',
-            contains("Field 'backfillResponseHandler' has not been initialized"),
+            contains(
+              "Field 'backfillResponseHandler' has not been initialized",
+            ),
           ),
         ),
       );
@@ -3461,12 +1091,14 @@ void main() {
         await expectLater(
           processor.process(event: event, journalDb: journalDb),
           throwsA(
-          isA<Error>().having(
-            (e) => e.toString(),
-            'toString',
-            contains("Field 'backfillResponseHandler' has not been initialized"),
+            isA<Error>().having(
+              (e) => e.toString(),
+              'toString',
+              contains(
+                "Field 'backfillResponseHandler' has not been initialized",
+              ),
+            ),
           ),
-        ),
         );
       },
     );
@@ -3738,401 +1370,6 @@ void main() {
     );
   });
 
-  group('EntryLink sequence log recording -', () {
-    late MockSyncSequenceLogService mockSequenceService;
-
-    setUp(() {
-      mockSequenceService = MockSyncSequenceLogService();
-    });
-
-    test(
-      'records entry link in sequence log when vectorClock and originatingHostId present',
-      () async {
-        const vc = VectorClock({'host-A': 5});
-        final link = EntryLink.basic(
-          id: 'seq-link-1',
-          fromId: 'from-1',
-          toId: 'to-1',
-          createdAt: DateTime(2024),
-          updatedAt: DateTime(2024),
-          vectorClock: vc,
-        );
-        final message = SyncMessage.entryLink(
-          entryLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-A',
-        );
-
-        when(
-          () => mockSequenceService.recordReceivedEntryLink(
-            linkId: any(named: 'linkId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-          ),
-        ).thenAnswer((_) async => []);
-
-        final processorWithSeq = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSequenceService,
-        );
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-        await processorWithSeq.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => mockSequenceService.recordReceivedEntryLink(
-            linkId: 'seq-link-1',
-            vectorClock: vc,
-            originatingHostId: 'host-A',
-          ),
-        ).called(1);
-      },
-    );
-
-    test(
-      'logs gap detection when recordReceivedEntryLink returns gaps',
-      () async {
-        const vc = VectorClock({'host-B': 10});
-        final link = EntryLink.basic(
-          id: 'seq-link-gaps',
-          fromId: 'from-gap',
-          toId: 'to-gap',
-          createdAt: DateTime(2024),
-          updatedAt: DateTime(2024),
-          vectorClock: vc,
-        );
-        final message = SyncMessage.entryLink(
-          entryLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-B',
-        );
-
-        when(
-          () => mockSequenceService.recordReceivedEntryLink(
-            linkId: any(named: 'linkId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-          ),
-        ).thenAnswer(
-          (_) async => [
-            (hostId: 'host-B', counter: 8),
-            (hostId: 'host-B', counter: 9),
-          ],
-        );
-
-        final processorWithSeq = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSequenceService,
-        );
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-        await processorWithSeq.process(event: event, journalDb: journalDb);
-
-        verify(
-          () => loggingService.captureEvent(
-            contains('apply.entryLink.gapsDetected count=2'),
-            domain: LogDomains.sync,
-            subDomain: 'processor.gapDetection',
-          ),
-        ).called(1);
-      },
-    );
-
-    test('handles recordReceivedEntryLink exceptions gracefully', () async {
-      const vc = VectorClock({'host-C': 3});
-      final link = EntryLink.basic(
-        id: 'seq-link-error',
-        fromId: 'from-err',
-        toId: 'to-err',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: vc,
-      );
-      final message = SyncMessage.entryLink(
-        entryLink: link,
-        status: SyncEntryStatus.update,
-        originatingHostId: 'host-C',
-      );
-
-      when(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: any(named: 'linkId'),
-          vectorClock: any(named: 'vectorClock'),
-          originatingHostId: any(named: 'originatingHostId'),
-        ),
-      ).thenThrow(Exception('sequence log error'));
-
-      final processorWithSeq = SyncEventProcessor(
-        loggingService: loggingService,
-        updateNotifications: updateNotifications,
-        aiConfigRepository: aiConfigRepository,
-        settingsDb: settingsDb,
-        journalEntityLoader: journalEntityLoader,
-        sequenceLogService: mockSequenceService,
-      );
-
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-      // Should not throw - errors are caught and logged
-      await processorWithSeq.process(event: event, journalDb: journalDb);
-
-      // Verify exception was logged
-      verify(
-        () => loggingService.captureException(
-          any<Object>(),
-          domain: 'SYNC_SEQUENCE',
-          subDomain: 'recordReceived',
-          stackTrace: any<StackTrace>(named: 'stackTrace'),
-        ),
-      ).called(1);
-    });
-
-    test('skips sequence log when vectorClock is null', () async {
-      final link = EntryLink.basic(
-        id: 'seq-link-no-vc',
-        fromId: 'from-novc',
-        toId: 'to-novc',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: null,
-      );
-      final message = SyncMessage.entryLink(
-        entryLink: link,
-        status: SyncEntryStatus.update,
-        originatingHostId: 'host-D',
-      );
-
-      final processorWithSeq = SyncEventProcessor(
-        loggingService: loggingService,
-        updateNotifications: updateNotifications,
-        aiConfigRepository: aiConfigRepository,
-        settingsDb: settingsDb,
-        journalEntityLoader: journalEntityLoader,
-        sequenceLogService: mockSequenceService,
-      );
-
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-      await processorWithSeq.process(event: event, journalDb: journalDb);
-
-      // Sequence log should NOT be called
-      verifyNever(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: any(named: 'linkId'),
-          vectorClock: any(named: 'vectorClock'),
-          originatingHostId: any(named: 'originatingHostId'),
-        ),
-      );
-    });
-
-    test('skips sequence log when originatingHostId is null', () async {
-      const vc = VectorClock({'host-E': 1});
-      final link = EntryLink.basic(
-        id: 'seq-link-no-origin',
-        fromId: 'from-noorig',
-        toId: 'to-noorig',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: vc,
-      );
-      final message = SyncMessage.entryLink(
-        entryLink: link,
-        status: SyncEntryStatus.update,
-        // No originatingHostId
-      );
-
-      final processorWithSeq = SyncEventProcessor(
-        loggingService: loggingService,
-        updateNotifications: updateNotifications,
-        aiConfigRepository: aiConfigRepository,
-        settingsDb: settingsDb,
-        journalEntityLoader: journalEntityLoader,
-        sequenceLogService: mockSequenceService,
-      );
-
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-      await processorWithSeq.process(event: event, journalDb: journalDb);
-
-      // Sequence log should NOT be called
-      verifyNever(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: any(named: 'linkId'),
-          vectorClock: any(named: 'vectorClock'),
-          originatingHostId: any(named: 'originatingHostId'),
-        ),
-      );
-    });
-
-    test('records when rows=0 but link exists locally', () async {
-      const vc = VectorClock({'host-F': 7});
-      final link = EntryLink.basic(
-        id: 'seq-link-exists',
-        fromId: 'from-exists',
-        toId: 'to-exists',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: vc,
-      );
-      final message = SyncMessage.entryLink(
-        entryLink: link,
-        status: SyncEntryStatus.update,
-        originatingHostId: 'host-F',
-      );
-
-      when(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: any(named: 'linkId'),
-          vectorClock: any(named: 'vectorClock'),
-          originatingHostId: any(named: 'originatingHostId'),
-        ),
-      ).thenAnswer((_) async => []);
-
-      final processorWithSeq = SyncEventProcessor(
-        loggingService: loggingService,
-        updateNotifications: updateNotifications,
-        aiConfigRepository: aiConfigRepository,
-        settingsDb: settingsDb,
-        journalEntityLoader: journalEntityLoader,
-        sequenceLogService: mockSequenceService,
-      );
-
-      when(() => event.text).thenReturn(encodeMessage(message));
-      // rows=0 (no-op upsert)
-      when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 0);
-      // But link exists locally
-      when(
-        () => journalDb.entryLinkById('seq-link-exists'),
-      ).thenAnswer((_) async => link);
-
-      await processorWithSeq.process(event: event, journalDb: journalDb);
-
-      // Sequence log SHOULD be called because link exists
-      verify(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: 'seq-link-exists',
-          vectorClock: vc,
-          originatingHostId: 'host-F',
-        ),
-      ).called(1);
-    });
-
-    test(
-      'skips recording when rows=0 and link does not exist locally',
-      () async {
-        const vc = VectorClock({'host-G': 2});
-        final link = EntryLink.basic(
-          id: 'seq-link-missing',
-          fromId: 'from-missing',
-          toId: 'to-missing',
-          createdAt: DateTime(2024),
-          updatedAt: DateTime(2024),
-          vectorClock: vc,
-        );
-        final message = SyncMessage.entryLink(
-          entryLink: link,
-          status: SyncEntryStatus.update,
-          originatingHostId: 'host-G',
-        );
-
-        final processorWithSeq = SyncEventProcessor(
-          loggingService: loggingService,
-          updateNotifications: updateNotifications,
-          aiConfigRepository: aiConfigRepository,
-          settingsDb: settingsDb,
-          journalEntityLoader: journalEntityLoader,
-          sequenceLogService: mockSequenceService,
-        );
-
-        when(() => event.text).thenReturn(encodeMessage(message));
-        // rows=0 (no-op upsert)
-        when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 0);
-        // Link does NOT exist locally
-        when(
-          () => journalDb.entryLinkById('seq-link-missing'),
-        ).thenAnswer((_) async => null);
-
-        await processorWithSeq.process(event: event, journalDb: journalDb);
-
-        // Sequence log should NOT be called
-        verifyNever(
-          () => mockSequenceService.recordReceivedEntryLink(
-            linkId: any(named: 'linkId'),
-            vectorClock: any(named: 'vectorClock'),
-            originatingHostId: any(named: 'originatingHostId'),
-          ),
-        );
-      },
-    );
-
-    test('passes coveredVectorClocks to recordReceivedEntryLink', () async {
-      const vc = VectorClock({'host-A': 5});
-      const coveredClock1 = VectorClock({'host-A': 3});
-      const coveredClock2 = VectorClock({'host-A': 4});
-      final link = EntryLink.basic(
-        id: 'seq-link-covered',
-        fromId: 'from-covered',
-        toId: 'to-covered',
-        createdAt: DateTime(2024),
-        updatedAt: DateTime(2024),
-        vectorClock: vc,
-      );
-      final message = SyncMessage.entryLink(
-        entryLink: link,
-        status: SyncEntryStatus.update,
-        originatingHostId: 'host-A',
-        coveredVectorClocks: [coveredClock1, coveredClock2],
-      );
-
-      when(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: any(named: 'linkId'),
-          vectorClock: any(named: 'vectorClock'),
-          originatingHostId: any(named: 'originatingHostId'),
-          coveredVectorClocks: any(named: 'coveredVectorClocks'),
-        ),
-      ).thenAnswer((_) async => []);
-
-      final processorWithSeq = SyncEventProcessor(
-        loggingService: loggingService,
-        updateNotifications: updateNotifications,
-        aiConfigRepository: aiConfigRepository,
-        settingsDb: settingsDb,
-        journalEntityLoader: journalEntityLoader,
-        sequenceLogService: mockSequenceService,
-      );
-
-      when(() => event.text).thenReturn(encodeMessage(message));
-      when(() => journalDb.upsertEntryLink(any())).thenAnswer((_) async => 1);
-
-      await processorWithSeq.process(event: event, journalDb: journalDb);
-
-      verify(
-        () => mockSequenceService.recordReceivedEntryLink(
-          linkId: 'seq-link-covered',
-          vectorClock: vc,
-          originatingHostId: 'host-A',
-          coveredVectorClocks: [coveredClock1, coveredClock2],
-        ),
-      ).called(1);
-    });
-  });
-
   group('runWithDeferredMissingEntryNudges', () {
     test('falls back to plain action for mock processor instances', () async {
       final mockProcessor = MockSyncEventProcessor();
@@ -4247,486 +1484,6 @@ void main() {
         );
       },
     );
-  });
-
-  group('SyncEventProcessor - SyncOutboxBundle integration', () {
-    test(
-      'an inline outboxBundle is wired into the apply pipeline and each '
-      'child flows through its existing per-type handler. Detailed '
-      'unpacker behaviour (sidecar resolution, fault isolation, nested '
-      'bundles, ordering) is covered in '
-      'test/features/sync/matrix/outbox_bundle_unpacker_test.dart.',
-      () async {
-        const bundle = SyncOutboxBundle(
-          children: [
-            SyncMessage.aiConfigDelete(id: 'cfg-1'),
-            SyncMessage.aiConfigDelete(id: 'cfg-2'),
-          ],
-        );
-        when(() => event.text).thenReturn(encodeMessage(bundle));
-
-        await processor.process(event: event, journalDb: journalDb);
-
-        final captured = verify(
-          () => aiConfigRepository.deleteConfig(
-            captureAny<String>(),
-            fromSync: true,
-          ),
-        ).captured;
-        expect(captured, ['cfg-1', 'cfg-2']);
-      },
-    );
-
-    test(
-      'resolveOutboxBundleManifestForTesting returns null for a null '
-      'jsonPath — proves the manifest resolver early-skips before any disk '
-      'IO when the bundle stub has no jsonPath to fetch',
-      () async {
-        final result = await processor.resolveOutboxBundleManifestForTesting(
-          null,
-        );
-        expect(result, isNull);
-      },
-    );
-
-    group('manifest resolver on disk', () {
-      late Directory tempDir;
-
-      setUp(() {
-        tempDir = Directory.systemTemp.createTempSync('outbox_bundle_test');
-        if (getIt.isRegistered<Directory>()) {
-          getIt.unregister<Directory>();
-        }
-        getIt.registerSingleton<Directory>(tempDir);
-      });
-
-      tearDown(() async {
-        await getIt.reset();
-        if (tempDir.existsSync()) {
-          tempDir.deleteSync(recursive: true);
-        }
-      });
-
-      test(
-        'falls back to a disk read when no descriptor is registered, '
-        'parses the v1 manifest, materializes each SyncJournalEntity '
-        "child's payload to its declared jsonPath, and returns a "
-        'reconstructed bundle whose children round-trip through '
-        'SyncMessage.fromJson',
-        () async {
-          final entity = JournalEntry(
-            meta: Metadata(
-              id: 'bundle-entry-1',
-              createdAt: DateTime.utc(2026, 4, 25),
-              updatedAt: DateTime.utc(2026, 4, 25),
-              dateFrom: DateTime.utc(2026, 4, 25),
-              dateTo: DateTime.utc(2026, 4, 25),
-              vectorClock: const VectorClock({'host-A': 3}),
-            ),
-            entryText: const EntryText(plainText: 'inside the bundle'),
-          );
-
-          const entityRelativePath =
-              '/journal/2026-04-25/bundle-entry-1.entry.json';
-          const bundleRelativePath = '/outbox_bundles/test-bundle.json';
-
-          final manifest = <String, dynamic>{
-            'version': 1,
-            'entries': [
-              <String, dynamic>{
-                'envelope': const SyncMessage.journalEntity(
-                  id: 'bundle-entry-1',
-                  jsonPath: entityRelativePath,
-                  vectorClock: VectorClock({'host-A': 3}),
-                  status: SyncEntryStatus.update,
-                ).toJson(),
-                'payload': entity.toJson(),
-              },
-              <String, dynamic>{
-                'envelope': const SyncMessage.aiConfigDelete(
-                  id: 'cfg-1',
-                ).toJson(),
-              },
-            ],
-          };
-
-          // Drop the manifest at the disk-fallback location: the resolver
-          // will read it via targetFile.readAsString() because no
-          // descriptor event is registered.
-          final manifestFile = File(
-            path.join(
-              tempDir.path,
-              stripLeadingSlashes(bundleRelativePath),
-            ),
-          )..parent.createSync(recursive: true);
-          manifestFile.writeAsStringSync(json.encode(manifest));
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNotNull);
-          expect(resolved!.jsonPath, bundleRelativePath);
-          expect(resolved.children, hasLength(2));
-          expect(resolved.children.first, isA<SyncJournalEntity>());
-          expect(resolved.children.last, isA<SyncAiConfigDelete>());
-
-          // The journal entity payload was materialized to its declared
-          // jsonPath under the documents directory, so the apply pipeline's
-          // SmartJournalEntityLoader reads it locally as a cache hit.
-          final materialized = File(
-            path.join(tempDir.path, stripLeadingSlashes(entityRelativePath)),
-          );
-          expect(materialized.existsSync(), isTrue);
-          final roundTripped = JournalEntity.fromJson(
-            json.decode(materialized.readAsStringSync())
-                as Map<String, dynamic>,
-          );
-          expect(roundTripped.meta.id, 'bundle-entry-1');
-          expect(
-            roundTripped.meta.vectorClock,
-            const VectorClock({'host-A': 3}),
-          );
-        },
-      );
-
-      test(
-        'returns null when the manifest version does not match — receivers '
-        'reject unknown wire shapes and let the surrounding retry mechanics '
-        're-deliver the bundle under a forward-compatible code path instead '
-        'of feeding garbage into JournalEntity.fromJson',
-        () async {
-          const bundleRelativePath = '/outbox_bundles/v99.json';
-          final manifestFile = File(
-            path.join(
-              tempDir.path,
-              stripLeadingSlashes(bundleRelativePath),
-            ),
-          )..parent.createSync(recursive: true);
-          manifestFile.writeAsStringSync(
-            json.encode(<String, dynamic>{
-              'version': 99,
-              'entries': const <Map<String, dynamic>>[],
-            }),
-          );
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNull);
-          verify(
-            () => loggingService.captureException(
-              any<Object>(
-                that: isA<String>().having(
-                  (msg) => msg,
-                  'message',
-                  contains('manifest version=99 unsupported'),
-                ),
-              ),
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.unknownVersion',
-              stackTrace: any<StackTrace?>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'aborts the whole bundle (returns null) when a SyncJournalEntity '
-        'envelope has no inline payload — silently skipping the child '
-        'would let the bundle ack while the entity never reaches the '
-        'local DB; peers recover the missing entry via the sequence-log '
-        'backfill path',
-        () async {
-          const bundleRelativePath = '/outbox_bundles/missing-payload.json';
-          final manifestFile = File(
-            path.join(
-              tempDir.path,
-              stripLeadingSlashes(bundleRelativePath),
-            ),
-          )..parent.createSync(recursive: true);
-          manifestFile.writeAsStringSync(
-            json.encode(<String, dynamic>{
-              'version': 1,
-              'entries': [
-                <String, dynamic>{
-                  'envelope': const SyncMessage.journalEntity(
-                    id: 'orphan',
-                    jsonPath: '/journal/2026-04-25/orphan.entry.json',
-                    vectorClock: VectorClock({'host-A': 1}),
-                    status: SyncEntryStatus.update,
-                  ).toJson(),
-                  // No `payload` key — this is the malformed case.
-                },
-              ],
-            }),
-          );
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNull);
-          verify(
-            () => loggingService.captureException(
-              any<Object>(
-                that: isA<String>().having(
-                  (msg) => msg,
-                  'message',
-                  contains('missing payload for SyncJournalEntity'),
-                ),
-              ),
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.missingPayload',
-              stackTrace: any<StackTrace?>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'when local DB version dominates the incoming envelope, refreshes '
-        'the on-disk cache from local.toJson() rather than overwriting it '
-        'with a staler bundled payload — guarantees SmartJournalEntityLoader '
-        'never serves a stale fixture even when the cache file was missing '
-        'or out-of-date',
-        () async {
-          final localEntity = JournalEntry(
-            meta: Metadata(
-              id: 'dominant-id',
-              createdAt: DateTime.utc(2026, 4, 25),
-              updatedAt: DateTime.utc(2026, 4, 25),
-              dateFrom: DateTime.utc(2026, 4, 25),
-              dateTo: DateTime.utc(2026, 4, 25),
-              vectorClock: const VectorClock({'host-A': 9}),
-            ),
-            entryText: const EntryText(plainText: 'canonical local'),
-          );
-
-          when(
-            () => journalDb.journalEntityMapForIds(any<Iterable<String>>()),
-          ).thenAnswer((_) async => {'dominant-id': localEntity});
-
-          final processorWithDb = SyncEventProcessor(
-            loggingService: loggingService,
-            updateNotifications: updateNotifications,
-            aiConfigRepository: aiConfigRepository,
-            settingsDb: settingsDb,
-            journalEntityLoader: journalEntityLoader,
-            journalDb: journalDb,
-          );
-
-          const entityRelativePath =
-              '/journal/2026-04-25/dominant-id.entry.json';
-          const bundleRelativePath = '/outbox_bundles/dominates.json';
-
-          // Pre-populate the cache file with an OLDER version to make the
-          // refresh observable: if the resolver did the right thing, the
-          // post-resolve disk content equals localEntity.toJson(); if it
-          // did the wrong thing, it stays as the older fixture.
-          final cacheFile = File(
-            path.join(tempDir.path, stripLeadingSlashes(entityRelativePath)),
-          )..parent.createSync(recursive: true);
-          cacheFile.writeAsStringSync(
-            json.encode(<String, dynamic>{'stale': true}),
-          );
-
-          final manifestFile = File(
-            path.join(
-              tempDir.path,
-              stripLeadingSlashes(bundleRelativePath),
-            ),
-          )..parent.createSync(recursive: true);
-          manifestFile.writeAsStringSync(
-            json.encode(<String, dynamic>{
-              'version': 1,
-              'entries': [
-                <String, dynamic>{
-                  'envelope': const SyncMessage.journalEntity(
-                    id: 'dominant-id',
-                    jsonPath: entityRelativePath,
-                    // Older than localEntity's VC -> local dominates.
-                    vectorClock: VectorClock({'host-A': 1}),
-                    status: SyncEntryStatus.update,
-                  ).toJson(),
-                  'payload': <String, dynamic>{'should': 'not-be-written'},
-                },
-              ],
-            }),
-          );
-
-          final resolved = await processorWithDb
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNotNull);
-          expect(resolved!.children.single, isA<SyncJournalEntity>());
-
-          final refreshed = JournalEntity.fromJson(
-            json.decode(cacheFile.readAsStringSync()) as Map<String, dynamic>,
-          );
-          // Cache now reflects the canonical DB version, not the bundled
-          // one and not the stale pre-test fixture.
-          expect(refreshed.meta.id, 'dominant-id');
-          expect(
-            refreshed.meta.vectorClock,
-            const VectorClock({'host-A': 9}),
-          );
-        },
-      );
-
-      test(
-        'aborts the whole bundle when an envelope.jsonPath escapes the '
-        'documents sandbox — defence in depth against tampered manifests',
-        () async {
-          const bundleRelativePath = '/outbox_bundles/bad-path.json';
-          final manifestFile = File(
-            path.join(
-              tempDir.path,
-              stripLeadingSlashes(bundleRelativePath),
-            ),
-          )..parent.createSync(recursive: true);
-          manifestFile.writeAsStringSync(
-            json.encode(<String, dynamic>{
-              'version': 1,
-              'entries': [
-                <String, dynamic>{
-                  'envelope': const SyncMessage.journalEntity(
-                    id: 'evil',
-                    // Multi-level traversal escapes the documents dir;
-                    // resolveJsonCandidateFile throws FileSystemException
-                    // (a single `/..` collapses back to `/` under
-                    // `path.normalize`, which would not escape).
-                    jsonPath: '../../escape.entry.json',
-                    vectorClock: VectorClock({'host-A': 1}),
-                    status: SyncEntryStatus.update,
-                  ).toJson(),
-                  'payload': <String, dynamic>{'meta': <String, dynamic>{}},
-                },
-              ],
-            }),
-          );
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNull);
-          verify(
-            () => loggingService.captureException(
-              any<Object>(that: isA<FileSystemException>()),
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.invalidEntryPath',
-              stackTrace: any<StackTrace?>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'returns null on a malformed manifest payload (missing or '
-        'non-list `entries` field) — the malformed shape never reaches '
-        'the per-child loop and surrounding diagnostics fire on a '
-        'distinct subDomain so the failure is greppable',
-        () async {
-          const bundleRelativePath = '/outbox_bundles/no-entries.json';
-          File(
-              path.join(
-                tempDir.path,
-                stripLeadingSlashes(bundleRelativePath),
-              ),
-            )
-            ..parent.createSync(recursive: true)
-            ..writeAsStringSync(
-              json.encode(<String, dynamic>{'version': 1}),
-            );
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNull);
-          verify(
-            () => loggingService.captureException(
-              any<Object>(
-                that: isA<String>().having(
-                  (msg) => msg,
-                  'message',
-                  contains('manifest missing entries array'),
-                ),
-              ),
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.malformed',
-              stackTrace: any<StackTrace?>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'silently skips malformed manifest entries (non-Map shapes, '
-        'envelope JSON that does not deserialize, nested SyncOutboxBundle '
-        'children) and resolves the rest — partial progress is preferable '
-        'to dropping the whole bundle when the rotten entry is the only '
-        'thing wrong',
-        () async {
-          const bundleRelativePath = '/outbox_bundles/mixed.json';
-          File(
-              path.join(
-                tempDir.path,
-                stripLeadingSlashes(bundleRelativePath),
-              ),
-            )
-            ..parent.createSync(recursive: true)
-            ..writeAsStringSync(
-              json.encode(<String, dynamic>{
-                'version': 1,
-                'entries': [
-                  // Not a map → skipped silently.
-                  'not-a-map',
-                  // Map but envelope is not a Map → skipped silently.
-                  <String, dynamic>{'envelope': 'not-a-map'},
-                  // Envelope JSON that fails SyncMessage.fromJson →
-                  // captured as envelopeParse but resolution continues.
-                  <String, dynamic>{
-                    'envelope': <String, dynamic>{
-                      'runtimeType': 'no-such-variant',
-                    },
-                  },
-                  // Nested SyncOutboxBundle child → defensively skipped
-                  // by the resolver before the unpacker even sees it.
-                  <String, dynamic>{
-                    'envelope': const SyncOutboxBundle(
-                      children: [],
-                    ).toJson(),
-                  },
-                  // A valid inline-only child rounds out the manifest so
-                  // the resolver still produces a usable bundle instead
-                  // of returning null.
-                  <String, dynamic>{
-                    'envelope': const SyncMessage.aiConfigDelete(
-                      id: 'cfg-good',
-                    ).toJson(),
-                  },
-                ],
-              }),
-            );
-
-          final resolved = await processor
-              .resolveOutboxBundleManifestForTesting(bundleRelativePath);
-
-          expect(resolved, isNotNull);
-          expect(resolved!.children, hasLength(1));
-          expect(resolved.children.single, isA<SyncAiConfigDelete>());
-          // The envelope-parse error path captured an exception under
-          // its own subDomain — at least once, exactly for the rotten
-          // runtimeType entry above.
-          verify(
-            () => loggingService.captureException(
-              any<Object>(),
-              domain: 'MATRIX_SERVICE',
-              subDomain: 'processor.resolve.outboxBundle.envelopeParse',
-              stackTrace: any<StackTrace?>(named: 'stackTrace'),
-            ),
-          ).called(1);
-        },
-      );
-    });
   });
 
   group('SyncEventProcessor - self-echo skip', () {
